Use activity starter instead of result contract
